[PART-5]: Acceptance testing

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,16 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, only: %i[new create edit update destroy]
   before_action :set_question, only: %i[show edit update destroy]
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answers = @question.answers
+    @answer = @question.answers.build
+  end
 
   def new
     @question = Question.new
@@ -15,9 +19,9 @@ class QuestionsController < ApplicationController
   def edit; end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.build(question_params)
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new, status: :unprocessable_entity
     end
@@ -32,6 +36,7 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
+    redirect_to questions_path unless current_user
     if @question.destroy
       redirect_to questions_path
     else
